openclaw/requirements.txt
contextual-client>=0.11.0
anthropic>=0.86.0
apify-client>=2.5.0
redis>=7.4.0
python-dotenv>=1.2.2
requests>=2.33.0
pdfplumber>=0.11.0
pytesseract>=0.3.13
Pillow>=10.0.0
weasyprint>=68.0
python-telegram-bot>=22.0
